// Bottleneko.ScriptRuntime/neko/internal/api/bottleneko.gen.ts
export interface DiscordProtocolConfiguration {
    $type: 'Discord';
    token: string;
    receiveEvents: boolean;
    isPresenceIntentEnabled: boolean;
    isServerMembersIntentEnabled: boolean;
    isMessageContentIntentEnabled: boolean;
    proxyId: string | null;
}

export type ProtocolConfiguration
    = DiscordProtocolConfiguration
        | TelegramProtocolConfiguration
        | TwitchProtocolConfiguration;

export interface TelegramProtocolConfiguration {
    $type: 'Telegram';
    token: string;
    receiveEvents: boolean;
    proxyId: string | null;
}
// ...
export interface TwitchProtocolConfiguration {
    $type: 'Twitch';
    auth: TwitchAuth;
    receiveEvents: boolean;
    channels: TwitchProtocolChannel[];
    proxyId: string | null;
}
// ...
export type ConnectionStatus
    = 'NotConnected'
        | 'Connecting'
        | 'Connected'
        | 'Reconnecting'
        | 'DelayedReconnect'
        | 'Stopping'
        | 'Error';
export const ConnectionStatusValues : ConnectionStatus[] = [
    'NotConnected',
    'Connecting',
    'Connected',
    'Reconnecting',
    'DelayedReconnect',
    'Stopping',
    'Error',
];


export type Protocol
    = 'Discord'
        | 'Telegram'
        | 'Twitch';
export const ProtocolValues : Protocol[] = [
    'Discord',
    'Telegram',
    'Twitch',
];

export interface ExtendedConnectionStatus {
    status: ConnectionStatus;
    statusChangeDelay: number;
}

export interface ConnectionDto {
    id: string;
    name: string;
    protocol: Protocol;
    autoStart: boolean;
    config: ProtocolConfiguration;
    extendedStatus: ExtendedConnectionStatus;
}